report/pyproject.toml
[project]
name = "campaign-supplementary-report-to-sensor"
version = "0.1.0"
description = ""
authors = [
    {name = "fushengbao",email = "fushengbao@ihuayue.cn"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "mysql-connector (>=2.2.9,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
